+++ anki_miner/interfaces/__init__.py
@@ -4,5 +4,12 @@ from .dictionary_provider import DictionaryProvider
 from .expression_audio import ExpressionAudioFetcher
 from .presenter import PresenterProtocol
 from .progress import ProgressCallback
+from .sentence_audio import SentenceAudioFetcher
 
-__all__ = ["DictionaryProvider", "ExpressionAudioFetcher", "PresenterProtocol", "ProgressCallback"]
+__all__ = [
+    "DictionaryProvider",
+    "ExpressionAudioFetcher",
+    "PresenterProtocol",
+    "ProgressCallback",
+    "SentenceAudioFetcher",
+]